refactor: toposort.ts now used and exports isADependentOnB
