+++ .gdb_history
@@ -1,0 +1,4 @@
+bt
+disas
+r
+bt
