+++ src/helper/toposort.ts
@@ -15,37 +15,36 @@
  */
 
 import { Paul } from "@/paul";
-import type { CryptOpt } from "@/types";
+import type { Nodes, CryptOpt } from "@/types";
 
-import { assertStringArguments, assertStringNames, matchXD } from "./lamdas";
+import { assertStringArguments, assertStringNames, makeU64NameLimbs } from "./lamdas";
 
-export function toposort(_nodes: Readonly<Readonly<CryptOpt.StringOperation>[]>): number[] {
+export function toposort(
+  _nodes: Readonly<Readonly<CryptOpt.StringOperation>[]>,
+  neededBy: Map<string, Set<string>>,
+): number[] {
   _nodes.forEach(assertStringNames);
   _nodes.forEach(assertStringArguments);
   const nodes = _nodes as CryptOpt.StringOperation[];
   const sorted = [] as number[];
+  const unsorted = new Set(nodes.keys());
+
   let l = 0;
 
-  const availableXDD = new Set<string>();
-
   while (l < nodes.length) {
-    const canScheduleNow = nodes.reduce((acc, node, idx) => {
-      // if it is already in the sorted, then we cant do it again
-      if (sorted.includes(idx)) {
-        return acc;
+    const canScheduleNow = [] as number[];
+    unsorted.forEach((index) => {
+      let isIndexDependentOnAnyOfTheRest = false;
+      unsorted.forEach((rest) => {
+        if (isADependentOnB(index, rest, nodes, neededBy)) {
+          isIndexDependentOnAnyOfTheRest = true;
+          return;
+        }
+      });
+      if (!isIndexDependentOnAnyOfTheRest) {
+        canScheduleNow.push(index);
       }
-      const xdArgs = node.arguments.filter(matchXD);
-      if (
-        // if it has no xDD arguments
-        xdArgs.length === 0 ||
-        // or all of them are available
-        xdArgs.every((arg) => availableXDD.has(arg))
-      ) {
-        // you can schedule a node
-        acc.push(idx);
-      }
-      return acc;
-    }, new Array<number>());
+    });
 
     if (canScheduleNow.length === 0 && l < nodes.length) {
       const unsorted = nodes
@@ -54,15 +53,45 @@ export function toposort(_nodes: Readonly<Readonly<CryptOpt.StringOperation>[]>)
         .join(", ");
       throw new Error(`i cant schedule anything, but I want need to sort in ... ${unsorted} `);
     }
+
     // then pick a node from the available at random
     const choice = Paul.pick(canScheduleNow);
+
     // add it to sorted
     l = sorted.push(choice);
 
-    // and as the name is not available, add them
-    nodes[choice].name.forEach((name) => availableXDD.add(name));
-    // makeU64NameLimbs(nodes[choice]).forEach((name) => availableXDD.add(name));
+    // remove from unsorted
+    unsorted.delete(choice);
   }
 
   return sorted;
+}
+
+export function isADependentOnB(
+  a: number,
+  b: number,
+  nodes: Nodes,
+  neededBy: Map<string, Set<string>>,
+): boolean {
+  if (isNaN(a) || isNaN(b)) {
+    throw new Error("only support number numbers");
+  }
+  const A = nodes[a];
+  const B = nodes[b];
+  if (typeof A === "undefined" || typeof B === "undefined") {
+    throw new Error("only supports indexes, which result in non-undefined nodes. ");
+  }
+  const setOfDeps = makeU64NameLimbs(B).reduce((acc: Set<string>, limb) => {
+    const set = neededBy.get(limb);
+    if (set) {
+      set.forEach((s) => acc.add(s));
+    }
+    return acc;
+  }, new Set<string>());
+  // setOfDeps may be empty (if no one wants B... like outN[n])
+
+  // in the setOfDeps we only have the limbs
+  const needles = makeU64NameLimbs(A);
+  // needles are the 64-bit versions like x1,x2 or x100_1
+  return needles.some((n) => setOfDeps.has(n));
 }
