+++ test/model.helper.ts
@@ -17,7 +17,7 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 import { describe, expect, it } from "vitest";
 
-import { createDependencyRelation, isADependentOnB, nodeLookupMap } from "@/model";
+import { createDependencyRelation, nodeLookupMap } from "@/model";
 
 import { createModelHelpers } from "./test-helpers";
 
@@ -267,91 +267,6 @@ describe("model.helpers", () => {
         expect(n).toBeTruthy();
         expect(n!.size).toBe(1);
         expect(n!.has("x5")).toBe(true);
-      });
-    });
-    describe("node Dependence: 'isADependentOnB'", () => {
-      it("is x1 dependent on x2", () => {
-        const a = nodes.findIndex((n) => n.name[0] === "x1");
-        const b = nodes.findIndex((n) => n.name[0] === "x2");
-        expect(isADependentOnB(a, b, nodes, neededBy)).toBe(false);
-      });
-      it("is x3 dependent on x2", () => {
-        const a = nodes.findIndex((n) => n.name[0] === "x3");
-        const b = nodes.findIndex((n) => n.name[0] === "x2");
-        expect(isADependentOnB(a, b, nodes, neededBy)).toBe(true);
-      });
-      it("is x3 dependent on x1", () => {
-        const a = nodes.findIndex((n) => n.name[0] === "x3");
-        const b = nodes.findIndex((n) => n.name[0] === "x1");
-        expect(isADependentOnB(a, b, nodes, neededBy)).toBe(true);
-      });
-      it("is x2 dependent on x3", () => {
-        const a = nodes.findIndex((n) => n.name[0] === "x2");
-        const b = nodes.findIndex((n) => n.name[0] === "x3");
-        expect(isADependentOnB(a, b, nodes, neededBy)).toBe(false);
-      });
-      it("is x1 dependent on x100", () => {
-        const a = nodes.findIndex((n) => n.name[0] === "x1");
-        const b = nodes.findIndex((n) => n.name[0] === "x100");
-        expect(isADependentOnB(a, b, nodes, neededBy)).toBe(false);
-      });
-      it("is x100 dependent on x1", () => {
-        const a = nodes.findIndex((n) => n.name[0] === "x100");
-        const b = nodes.findIndex((n) => n.name[0] === "x1");
-        expect(isADependentOnB(a, b, nodes, neededBy)).toBe(true);
-      });
-      it("is anything dependent on out1[0]", () => {
-        const b = nodes.findIndex((n) => n.name[0] === "out1[0]");
-        nodes.forEach((n, i) => {
-          if (n.name[0] === "out1[0]") return;
-          expect(isADependentOnB(i, b, nodes, neededBy)).toBe(false);
-        });
-      });
-      it("is x100 dependent on x2", () => {
-        const a = nodes.findIndex((n) => n.name[0] === "x100");
-        const b = nodes.findIndex((n) => n.name[0] === "x2");
-        expect(isADependentOnB(a, b, nodes, neededBy)).toBe(true);
-      });
-      it("is x100 dependent on x3", () => {
-        const a = nodes.findIndex((n) => n.name[0] === "x100");
-        const b = nodes.findIndex((n) => n.name[0] === "x3");
-        expect(isADependentOnB(a, b, nodes, neededBy)).toBe(false);
-      });
-      it("is x4 dependent on 0xffffffffffffffff, ... not really supported, cuz imms are not nodes", () => {
-        const a = nodes.findIndex((n) => n.name[0] === "x4");
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        const b = nodes.findIndex((n) => n.name[0] === ("0xffffffffffffffff" as any));
-        expect(b).toBe(-1);
-        expect(() => isADependentOnB(a, b, nodes, neededBy)).toThrow();
-      });
-      it("is x100_1 dependent (shall error cuz x100_1 is no node)", () => {
-        const a = nodes.findIndex((n) => n.name[0] === "x100_1"); // a shall be undefined
-        expect(a).toBe(-1);
-        const b = nodes.findIndex((n) => n.name[0] === "x3");
-        expect(() => isADependentOnB(a, b, nodes, neededBy)).toThrow();
-      });
-      it("error if NaNs are provided", () => {
-        expect(() => isADependentOnB(Number.NaN, Number.NaN, nodes, neededBy)).toThrow();
-      });
-      it("is x27 dependent on x8", () => {
-        const a = nodes.findIndex((n) => n.name[0] === "x27");
-        const b = nodes.findIndex((n) => n.name[0] === "x8");
-        expect(isADependentOnB(a, b, nodes, neededBy)).toBe(true);
-      });
-      it("is x8 dependent on x27", () => {
-        const a = nodes.findIndex((n) => n.name[0] === "x8");
-        const b = nodes.findIndex((n) => n.name[0] === "x27");
-        expect(isADependentOnB(a, b, nodes, neededBy)).toBe(false);
-      });
-      it("is x28 dependent on x27", () => {
-        const a = nodes.findIndex((n) => n.name[0] === "x28");
-        const b = nodes.findIndex((n) => n.name[0] === "x27");
-        expect(isADependentOnB(a, b, nodes, neededBy)).toBe(true);
-      });
-      it("is x27 dependent on x28", () => {
-        const a = nodes.findIndex((n) => n.name[0] === "x27");
-        const b = nodes.findIndex((n) => n.name[0] === "x28");
-        expect(isADependentOnB(a, b, nodes, neededBy)).toBe(false);
       });
     });
   });
